# docs/swagger.yaml
basePath: /api
definitions:
  admin.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  admin.SignInResponse:
    properties:
      error:
        type: string
      status:
        type: string
      token:
        type: string
    type: object
  admin.SignUpRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  cart.AddRequest:
    properties:
      count:
        type: integer
      product_id:
        type: integer
    type: object
...
  customer.SignInRequest:
    properties:
      alias:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
  customer.SignInResponse:
    properties:
      error:
        type: string
      status:
        type: string
      token:
        type: string
    type: object
  customer.SignUpRequest:
    properties:
      alias:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
...
  order.GetResponse:
    properties:
      error:
        type: string
      orders:
        items:
          $ref: '#/definitions/storage.Order'
        type: array
      status:
        type: string
    type: object
...
  response.Response:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
...
  storage.Order:
    properties:
      date_time:
        type: string
      order_items:
        items:
          $ref: '#/definitions/storage.OrderItem'
        type: array
    type: object
  storage.OrderItem:
    properties:
      count:
        type: integer
      id:
        type: integer
      order_id:
        type: integer
      product:
        $ref: '#/definitions/storage.ProductInfo'
    type: object
  storage.ProductInfo:
    properties:
      description:
        type: string
      id:
        type: integer
      image_id:
        type: integer
      name:
        type: string
      price:
        type: integer
      website_id:
        type: integer
    type: object
...
host: localhost:8082
info:
  contact: {}
  title: ForeignKey
  version: "1.0"
paths:
  /admin/sign-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: sign in
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/admin.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.SignInResponse'
      summary: SingIn admin
      tags:
      - auth
  /admin/sign-up:
    post:
      consumes:
      - application/json
      parameters:
      - description: sign up
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/admin.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: SingUp admin
      tags:
      - auth
  /cart/add:
    post:
      consumes:
      - application/json
      description: Добавляет товар в корзину, если товар уже в корзине, увеличивает
        количество
      parameters:
      - description: product id and count
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/cart.AddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Create cart item
      tags:
      - cart
...
  /customer/sign-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: sign in
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/customer.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.SignInResponse'
      summary: SingIn customer
      tags:
      - auth
  /customer/sign-up:
    post:
      consumes:
      - application/json
      parameters:
      - description: sign up
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/customer.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: SingUp customer
      tags:
      - auth
...
  /order/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.GetResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all orders
      tags:
      - order
  /order/make:
    post:
      description: Создает заказ исходя из корзины покупателя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Make order
      tags:
      - order
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"